[#207] Refactor views.py and start to remove Django deps so it's easier to use CoVE as a library
