+++ cove/lib/exceptions.py
@@ -1,0 +1,11 @@
+class CoveInputDataError(Exception):
+    """
+    An error that we think is due to the data input by the user, rather than a
+    bug in the application.
+    """
+    def __init__(self, context=None):
+        if context:
+            self.context = context
+
+
+
