+++ cove/lib/tools.py
@@ -1,0 +1,45 @@
+import strict_rfc3339
+import datetime
+
+
+def ignore_errors(f):
+    def ignore(json_data, ignore_errors=False):
+        if ignore_errors:
+            try:
+                return f(json_data)
+            except (KeyError, TypeError, IndexError, AttributeError):
+                return {}
+        else:
+            return f(json_data)
+    return ignore
+
+
+def to_list(item):
+    if isinstance(item, list):
+        return item
+    return [item]
+
+
+def get_no_exception(item, key, fallback):
+    try:
+        return item.get(key, fallback)
+    except AttributeError:
+        return fallback
+
+
+def update_docs(document_parent, counter):
+    count = 0
+    documents = document_parent.get('documents', [])
+    for document in documents:
+        count += 1
+        doc_type = document.get("documentType")
+        if doc_type:
+            counter.update([doc_type])
+    return count
+
+
+def datetime_or_date(instance):
+    result = strict_rfc3339.validate_rfc3339(instance)
+    if result:
+        return result
+    return datetime.datetime.strptime(instance, "%Y-%m-%d")
